created a designated routes folder
