+++ views/articles.hbs
@@ -9,5 +9,10 @@
 </button>
 
 <br>
+<br>
+<button id="artLanding">
+  <a href="/articles">Back to Articles</a></button>
+
+<br>
 
 <br>{{> footer}}
