+++ views/partials/footer.hbs
@@ -1,1 +1,3 @@
-<a href="/">HOME</a>
+<button id="homeButton">
+  <a href="/">HOME</a>
+</button>
