+++ views/products.hbs
@@ -10,4 +10,9 @@
 </button>
 
 <br>
+<br>
+<button id="prodLanding">
+  <a href="/products">Back to Products</a></button>
+
+<br>
 <br>{{> footer}}
